requirements.txt
Flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
